requirements.txt
google-cloud-firestore==2.16.1
boto3==1.34.140
python-dotenv==1.0.1
